requirements.txt
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
boto3==1.16.27
botocore==1.19.27
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
crispy-forms-materialize==0.2
cryptography==3.3.1
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap4==2.3.1
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-forms-bootstrap==3.1.0
django-materializecss-form==1.1.17
django-password-reset==2.0
django-storages==1.10.1
django-webpush==0.3.3
djangorestframework==3.12.2
fcm-django==0.3.7
gunicorn==20.0.4
http-ece==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.1.0
incremental==17.5.0
jmespath==0.10.0
Pillow==8.0.1
psycopg2-binary==2.8.6
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyfcm==1.4.9
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2020.4
pywebpush==1.9.4
requests==2.25.1
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
treq==20.9.0
Twisted==20.3.0
txFCM==0.0.1
urllib3==1.26.2
zipp==3.4.0
zope.interface==5.2.0
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
boto3==1.16.27
botocore==1.19.27
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
crispy-forms-materialize==0.2
cryptography==3.3.1
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap4==2.3.1
django-braces==1.14.0
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-datetimepicker==3.14
django-forms-bootstrap==3.1.0
django-ical==1.7.1
django-materializecss-form==1.1.17
django-model-utils==4.1.1
django-password-reset==2.0
django-recurrence==1.10.3
django-star-ratings==0.9.1
django-starfield==1.0.post1
django-storages==1.10.1
django-toolbox==0.2
django-webpush==0.3.3
djangorestframework==3.12.2
fcm-django==0.3.7
gunicorn==20.0.4
http-ece==1.1.0
hyperlink==20.0.1
icalendar==4.0.7
idna==2.10
importlib-metadata==2.1.0
incremental==17.5.0
jmespath==0.10.0
Pillow==8.0.1
psycopg2-binary==2.8.6
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyfcm==1.4.9
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2020.4
pywebpush==1.9.4
requests==2.25.1
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
treq==20.9.0
Twisted==20.3.0
txFCM==0.0.1
urllib3==1.26.2
vobject==0.9.6.1
zipp==3.4.0
zope.interface==5.2.0
stripe==2.56.0
